+++ packages/_private/angular-demo/src/app.module.ts
@@ -6,7 +6,6 @@ import {
   SynergyComponentsModule,
 } from '@synergy-design-system/angular';
 import { AppComponent } from './app.component';
-import { DemoFieldset } from './demofieldset/demofieldset.component';
 import { Home } from './home/home.component';
 import { DemoForm } from './demoform/demoform.component';
 import { DemoFormValidate } from './demoformvalidate/demoformvalidate.component';
@@ -17,14 +16,13 @@ import {
   SynNavItemRouterLinkDirective,
   SynNavItemRouterLinkActiveDirective,
 } from './directives';
+import { DemoFieldSetModule } from './modules/demofieldset.module';
 
 @NgModule({
   declarations: [
     Home,
     AppComponent,
-    DemoFieldset,
     DemoForm,
-    DemoFormValidate,
     ThemeSwitchComponent,
   ],
   imports: [
@@ -36,7 +34,9 @@ import {
     SynNavItemRouterLinkDirective,
     SynNavItemRouterLinkActiveDirective,
     SizeSwitchComponent,
-],
+    DemoFormValidate,
+    DemoFieldSetModule,
+  ],
   providers: [],
   bootstrap: [AppComponent],
 })