+++ packages/_private/angular-demo/src/demoform/demoform.component.ts
@@ -1,4 +1,4 @@
-import type { SynChangeEvent, SynRange } from '@synergy-design-system/components';
+import type { SynChangeEvent } from '@synergy-design-system/components';
 import { highlightOptionRenderer, serialize } from '@synergy-design-system/components';
 import { Component, ElementRef, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';