+++ packages/_private/angular-demo/src/modules/demofieldset.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { DemoFieldset } from '../demofieldset/demofieldset.component';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+  declarations: [DemoFieldset],
+  imports: [CommonModule],
+  exports: [DemoFieldset],
+})
+export class DemoFieldSetModule {}