+++ packages/components/scripts/jobs/angular.js
@@ -1,4 +1,5 @@
 import path from 'path';
+import { deleteAsync } from 'del';
 import * as jobs from './angular/index.js';
 import {
   createRunFormat,
@@ -22,21 +23,29 @@ export const runCreateAngularWrappers = async ({
 }) => {
   const metadata = await getManifestData(componentDistDir);
 
-  // Internal react package paths for usage in sub packages
-  const outDir = path.join(angularPackageDir, './src');
+  // Internal angular package paths for usage in sub packages
+  const outDir = path.join(angularPackageDir);
   const componentsDir = path.join(outDir, 'components');
   const modulesDir = path.join(outDir, 'modules');
   const directivesDir = path.join(outDir, 'directives');
 
   const distDir = path.join(angularPackageDir, './dist');
+  const indexFile = path.join(outDir, 'index.ts');
 
   await runAdjustPackageVersion('Angular: Adjusting angular package.json version field...')(componentPackageDir, angularPackageDir);
-  await createRunPrepare('Angular: Cleaning up artifacts...')(outDir, distDir, modulesDir, componentsDir, directivesDir);
+  await createRunPrepare('Angular: Cleaning up artifacts...')(distDir, modulesDir, componentsDir, directivesDir);
+  // Remove old index file
+  await deleteAsync(indexFile, { force: true });
+
   await jobs.runCreateComponents(metadata, componentsDir);
   await jobs.runCreateNgModule(metadata, modulesDir);
   await jobs.runCreateFormsModule(modulesDir);
   await jobs.runCreateValidatorDirectives(directivesDir);
   await jobs.runCreateExports(outDir);
-  await runFormat(outDir);
-  await jobs.runNgPackagr();
+  // Run format for all subfolders
+  await runFormat(componentsDir);
+  await runFormat(modulesDir);
+  await runFormat(directivesDir);
+  await jobs.runAngularBuild();
+  await jobs.runAdjustPackageExports('Angular: Adjusting angular package exports...')(distDir);
 };