+++ packages/components/scripts/jobs/angular/createExports.js
@@ -5,16 +5,16 @@ import { createHeader, job } from '../shared.js';
 const headerComment = createHeader('angular');
 
 /**
- * Creates the synergy angular module, located at packages/angular/src/modules/synergy.module.ts
+ * Creates the index.ts exports file, located at packages/angular/index.ts
  */
 export const runCreateExports = job('Angular: Creating exports file...', async (outDir) => {
   const outFile = path.join(outDir, 'index.ts');
   const output = `
 ${headerComment}
-export * from './components';
-export * from './directives/validators';
-export * from './modules/forms.module';
-export * from './modules/synergy.module'; 
+export * from '@synergy-design-system/angular/components';
+export * from '@synergy-design-system/angular/directives/validators';
+export * from '@synergy-design-system/angular/modules/forms';
+export * from '@synergy-design-system/angular/modules/synergy'; 
 `.trim();
 
   await fs.writeFile(outFile, `${output}\n`, 'utf8');