+++ packages/components/scripts/jobs/angular/index.js
@@ -3,4 +3,5 @@ export * from './createExports.js';
 export * from './createFormsModule.js';
 export * from './createNgModule.js';
 export * from './createValidatorDirectives.js';
-export * from './ngPackagr.js';
+export * from './build.js';
+export * from './shared.js';